+++ pkg/idp/idp_test.go
@@ -7,7 +7,9 @@ import (
 
 	"giantswarm/dex-operator/pkg/idp/provider"
 	"giantswarm/dex-operator/pkg/idp/provider/mockprovider"
+	"giantswarm/dex-operator/pkg/key"
 
+	corev1 "k8s.io/api/core/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 )
 
@@ -46,6 +48,79 @@ func TestCreateProviderApps(t *testing.T) {
 			_, err := s.CreateProviderApps(tc.appConfig, context.Background())
 			if err != nil {
 				t.Fatal(err)
+			}
+		})
+	}
+}
+
+func TestGetBaseDomain(t *testing.T) {
+	testCases := []struct {
+		name           string
+		data           map[string]string
+		expectedDomain string
+	}{
+		{
+			name: "case 0",
+			data: map[string]string{
+				key.ClusterValuesConfigMapKey: `
+					something: "12"
+					baseDomain: hello.io
+					somethingelse: "false"
+					object:
+					  yes: no
+					`,
+			},
+			expectedDomain: "hello.io",
+		},
+		{
+			name: "case 1",
+			data: map[string]string{
+				key.ClusterValuesConfigMapKey: `
+					something: "12"
+					somethingelse: "false"
+					object:
+					  yes: no
+					baseDomain: hi.goodday.hello.io
+					`,
+			},
+			expectedDomain: "hi.goodday.hello.io",
+		},
+		{
+			name: "case 2",
+			data: map[string]string{
+				key.ClusterValuesConfigMapKey: `
+					something: "12"
+					somethingelse: "false"
+					object:
+					  yes: no
+					`,
+			},
+			expectedDomain: "",
+		},
+		{
+			name: "case 3",
+			data: map[string]string{
+				key.ClusterValuesConfigMapKey: `
+				baseDomain: hi.goodday.hello.io
+					something: "12"
+					somethingelse: "false"
+					object:
+					  yes: no
+					  baseDomain: no.goodday.hello.io
+					`,
+			},
+			expectedDomain: "hi.goodday.hello.io",
+		},
+	}
+
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			cm := &corev1.ConfigMap{
+				Data: tc.data,
+			}
+			baseDomain := getBaseDomainFromClusterValues(cm)
+			if baseDomain != tc.expectedDomain {
+				t.Fatalf("Expected %v to be equal to %v", baseDomain, tc.expectedDomain)
 			}
 		})
 	}